# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs",
        "description": "Multi Quest API Documentation",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "eaf331834ce7ca6883e34c6a123c1c1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "14a901c030e00a717962d8cacfba0a1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/addresses/{genre_id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get a list of addresses of a genre",
                "operationId": "6e6aedb490b0565e25eb9ba6cfbfdc4e",
                "parameters": [
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Genre ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/address_letters/{genre_id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get a list of address letter of a genre",
                "operationId": "3fb33c5155f474647ceede27692e9956",
                "parameters": [
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Genre ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/address/{genre_id}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Get an exact address",
                "operationId": "fd419200d38d47652108be88f7f86db5",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "genre_id",
                        "in": "path",
                        "description": "Genre ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "letter",
                        "in": "query",
                        "description": "Address letter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "С"
                    },
                    {
                        "name": "number",
                        "in": "query",
                        "description": "Address number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "76"
                    },
                    {
                        "name": "instance_id",
                        "in": "query",
                        "description": "Case instance id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Get a list of genres",
                "operationId": "e40500843f5dd80e6292f4edd20a01a1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/questions/{case_id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get a list of questions of a case",
                "operationId": "c690bb4313f5f333d710b229f96c9181",
                "parameters": [
                    {
                        "name": "case_id",
                        "in": "path",
                        "description": "Case ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/checkTips": {
            "get": {
                "tags": [
                    "Tips"
                ],
                "summary": "Check if tips are available and get them",
                "operationId": "205301e192916749b31a10fe082b97d8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "instance_id",
                        "in": "query",
                        "description": "Case instance id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/answer": {
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Answer to question",
                "operationId": "d0b663d84677970c89fd5d18333f1e66",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answer": {
                                        "description": "Text of the answer",
                                        "type": "string",
                                        "example": "answer"
                                    },
                                    "question_id": {
                                        "description": "ID of a question",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "instance_id": {
                                        "description": "Case instance id",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "User can't answer twice to one question"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Addresses",
            "description": "Addresses"
        },
        {
            "name": "Genres",
            "description": "Genres"
        },
        {
            "name": "Questions",
            "description": "Questions"
        },
        {
            "name": "Tips",
            "description": "Tips"
        }
    ]
}